scraper/requirements.txt
aio_pika == 9.4.3
telethon == 1.37.0
sqlalchemy == 2.0.36
psycopg2-binary == 2.9.10
python-dotenv == 1.0.1